refactor(hook): 拆分 ApplicationHook 冗长逻辑并重构 RequestManager
将 ApplicationHook 中冗长的广播处理逻辑拆分至 ApplicationBroadcastDispatcher，
同时将应用恢复与离线状态协调逻辑提取至 ApplicationResumeCoordinator。

此外，重构了 RequestManager 的内部结构，将 RPC 日志限流、请求注册器、缓存
策略及 JSON 降级工厂等辅助组件迁移至独立的 RequestManagerSupport 文件，
